Cache user permissions for individual resources
Single requests (especially PATCH requests) on ACL-protected resources
can end up resulting in thousands of redundant ACL lookups.  This
appears due to a consequence of the modeshape value mappers, which end
up consulting ACLs for every inspection of a property (triple) on a
node.

As a workaround, a per-session cache is added to
AbstractRolesAuthorizationDelegate, which caches the result of
"hasPermission" for each set of permissions on a node.]

So this does not fix the "thousands of lookups", but does fix the
"thousands of reads from the modeshape database" by introducing a cache.